requirements.txt
amplify
python-dotenv
matplotlib
